requirements.txt
aiobotocore==2.9.0
aiofiles==23.2.1
aiohttp==3.9.1
azure-cognitiveservices-speech==1.38.0
google-cloud-speech==2.25.0
fastapi==0.108.0
fastembed==0.3.5
huggingface-hub==0.24.3
litellm==1.65.0
llama_index==0.10.65
llama-index-vector-stores-lancedb==0.1.7
numpy==1.26.1
openai>=1.10.0
plivo==4.47.0
pydantic==2.5.3
pydub==0.25.1
pymongo==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
tiktoken>=0.6.0
torchaudio==2.0.1
twilio==8.9.0
uvicorn==0.22.0
websockets==15.0.1
onnxruntime>=1.16.3
scipy==1.11.4
semantic-router==0.0.46
sentence-transformers==3.0.1
tokenizers==0.15.2
optimum[onnxruntime]
lancedb==0.12.0
pymongo==4.8.0
httpx[http2]>=0.25.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0